Set `enumsAsTypes: true` in codegen config and remove `futureProofEnums: true`
This setting changes all graphql enums to use unions instead, which is much better to work with in TS. By not using the future proof setting, the union will not enclude "%future added value", which causes weird behaveior, because TS think this is a valid value.

According to the docs  `futureProofEnums: true` is used when generating the schema at runtime - which we dont do.

FIY: The generated code for dpl-cms did not have any enums at this moment.