Split next options and headers properly in dpl-cms fetcher
And make use of the nextjs fetch cache when fetching dpl-cms
configuration.
I think we should make use of the nextjs cache with all dpl-cms related
data, because we can have a common strategy for caching and revalidation
of the cache. And we are going to introduce a revalidation
endpoint to invalidate various content.