Simplify WorkCard.tsx
In order to make the component have a single purpose we divide it into multiple components that each have their own responsibility. Introduces WorkCardWithCaption.tsx + we wrap the card in a link if it needs to be linking somewhere (instead of having the link as a part of the component.