Replace `process.env` with validated `getEnv()` or `getServerEnv()`
This also means that we can safely remove many of the conditionals, which would return an error when an expected env variable is not found - this is all handled by env.ts now.

It is a bit redundant to both validate envs with zod and with conditionals for the goConfig, and I think a refactor might be due in the future, so that goConfig and getEnv has less overlap.