Re-introduced `TEST_MODE` env variable
Env handling for Nextjs is still quite borked.
While setting `NODE_ENV=test` makes Nextjs use the env.test file while building. But when trying to read NODE_ENV it will either resolve as `development` or `production`.

This makes it very unreliable for checking if the application is running in a testing env. By creating our own `TEST_MODE` env we can correctly control when and how this value is resolved.